Enable centering the elements to visualise
To handle the many different ranges of X,Y coordinates possible, we
create a transformation matrix based on the max and min values found in
those axis. This matrix is produced at the start of the visualisation.

For those cases where the area of interest is taller than wider, we
apply a -90 degree rotation on the Z axis. We do so aiming to maximize
screen real state usage.