Fix orthogonal collisions with segments
When the EV was displacing perpendicular to the segment, the previous
version of the collision resolution was producing division by zero and
NaN values.

NaN values where caused by floating point/rounding errors while computing
the angle of incidence.
Div0 errors are handled as a special case where the whole distance to
compensate is applied to the component perpendicular to the segment.