Change the EV apoptosis to be probabilistic.
The chance for an EV to become apoptotic is defined as a probabilistic threshold.
This threshold is compared per second.

Update the value of the kinematic viscosity of water at 20C

Collision resolution between initial and default EVs.
The EV in default collision will correct its position based on the collision overlap.
Whereas, the initial EV will stay in the same position it had at the
end of the previous iteration.

Introduced independent function for computing the grag force with/without brownian motion

Increase the amount of detail in some comments.

The source point selection for secretory cells with 2 source points was improved
to fix selecting a sp belonging to the cell