Fix for EVs in inital state having repeated collisions
When an EV in default state collides with one in initial state, that in
initial state has its time in initial state extended and it's position fixed.
However, repeating this extension many times can lead to undefined behaviour.
Therefore, the extension is limited to 2x the original value for
'time in initial state'. After that, the EV in initial state becomes apoptotic.