Change id generation for secreted EVs
When new agents are introduced in the GPU and we request a new Id,
the id assigned depends on the sequence of block execution. Because
actual block schedulling is not user-controlled, we cannot ensure the
agents get the same ID in repeated runs.
To fix this, we keep a counter of the vesicles generated per cell. At
secretion, this counter is added to the source cell id to produce the
ID for the new EV.
Because long ints are not available on linux, we use an unsigned int
for the variable ID.
After assessing the max number of secretory cells in our simulations,
we determined it lays within 65536. Therefore, we dedicate the first
5 digits of the Id to hold the source EV id. Then, the EV counter in
the cell is multipllied times 1e6 and added.
Example, cell: 45322 will secrete the Ev # 1234. The Id would be:
12344532. For cell 1 secreting EV 1, the ID will be 100001.